Created framework project for code to get Project Rome tokens
Project Rome's non-Graph services don't work with MSAL-retrieved tokens, so needed to use ADAL for this. ADAL and MSAL in the same project didn't work, so this was my solution.

Once Rome updates their services, this project could be removed and the existing AuthenticationManager can be used to get tokens.